cli - cleanup, parsing helper functions
