+++ cli/src/modules/estimate.rs
@@ -1,6 +1,5 @@
 use crate::imports::*;
 use kaspa_wallet_core::tx::PaymentDestination;
-use kaspa_wallet_core::utils::*;
 
 #[derive(Default, Handler)]
 #[help("Estimate the fees for a transaction of a given amount")]
@@ -16,13 +15,11 @@ impl Estimate {
             return Err("Usage: estimate <amount> [<priority fee>]".into());
         }
 
-        let amount = argv.get(0).cloned().unwrap_or_default();
-        let amount_sompi = try_kaspa_str_to_sompi(amount)?.unwrap_or(0);
-        let priority_fee_sompi = try_map_kaspa_str_to_sompi_i64(argv.get(1).map(String::as_str))?.unwrap_or(0);
-
+        let amount_sompi = try_parse_required_nonzero_kaspa_as_sompi_u64(argv.get(0))?;
+        let priority_fee_sompi = try_parse_optional_kaspa_as_sompi_i64(argv.get(1))?.unwrap_or(0);
         let abortable = Abortable::default();
 
-        // just need any address for an estimate
+        // just use any address for an estimate (change address)
         let change_address = account.change_address().await?;
         let destination = PaymentDestination::PaymentOutputs(PaymentOutputs::try_from((change_address.clone(), amount_sompi))?);
         let estimate = account.estimate(destination, priority_fee_sompi.into(), None, &abortable).await?;
