+++ cli/src/modules/transfer.rs
@@ -1,5 +1,4 @@
 use crate::imports::*;
-use kaspa_wallet_core::utils::*;
 
 #[derive(Default, Handler)]
 #[help("Transfer funds between wallet accounts")]
@@ -16,23 +15,14 @@ impl Transfer {
         }
 
         let target_account = argv.get(0).unwrap();
-        let amount = argv.get(1).unwrap();
-        let amount_sompi = helpers::kas_str_to_sompi(amount)?.ok_or_else(|| Error::custom("Invalid amount"))?;
-        let priority_fee = argv.get(2).map(String::as_str);
-        let priority_fee_sompi = try_map_kaspa_str_to_sompi_i64(priority_fee)?.unwrap_or(0);
-
         let target_account = ctx.find_accounts_by_name_or_id(target_account).await?;
         if target_account.id == account.id {
             return Err("Cannot transfer to the same account".into());
         }
+        let amount_sompi = try_parse_required_nonzero_kaspa_as_sompi_u64(argv.get(1))?;
+        let priority_fee_sompi = try_parse_optional_kaspa_as_sompi_i64(argv.get(2))?.unwrap_or(0);
         let target_address = target_account.receive_address().await?;
-        let wallet_secret = Secret::new(ctx.term().ask(true, "Enter wallet password: ").await?.trim().as_bytes().to_vec());
-
-        let payment_secret = if ctx.wallet().is_account_key_encrypted(&account).await?.is_some_and(|f| f) {
-            Some(Secret::new(ctx.term().ask(true, "Enter payment password: ").await?.trim().as_bytes().to_vec()))
-        } else {
-            None
-        };
+        let (wallet_secret, payment_secret) = ctx.ask_wallet_secret(Some(&account)).await?;
 
         let abortable = Abortable::default();
         let outputs = PaymentOutputs::try_from((target_address.clone(), amount_sompi))?;
