+++ cli/src/wizards/mod.rs
@@ -1,2 +1,3 @@
 pub mod account;
+pub mod mnemonic;
 pub mod wallet;
